django-app/requirements.txt
Django==3.0.5
gunicorn==20.0.4
psycopg2-binary==2.8.3
numpy==1.18.1
scipy==1.4.1
opencv-python==4.1.2.30
pillow==5.0.0
pybind11==2.5.0
torch==1.4.0
torchvision==0.5.0
cryptography==2.8
Werkzeug==1.0.1
Celery==4.4.2
django-session-cleanup==3.0.0
django-cors-headers==3.2.1
djangorestframework==3.11.0
